`miio` connection layer: refactoring
